chore: moved users schema to user.py, changed imports to match user.py
