+++ Apps/pisowise-be/models/user.py
@@ -1,0 +1,17 @@
+from pydantic import BaseModel, EmailStr
+
+
+# User Create Model
+class UserCreate(BaseModel):
+    email: EmailStr
+    username: str
+
+
+# User Response Model
+class UserResponse(BaseModel):
+    user_id: str
+    email: EmailStr
+    username: str
+
+    class Config:
+        from_attributes = True
